requirements.txt
streamlit==1.32.0
google-generativeai==0.5.0
groq==0.5.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0